+++ CHANGELOG.md
@@ -10,7 +10,7 @@ New developments should use these parameters, to be compliant with default value
 
 Both mode (client and server):
 * The GraphQL schema can now be split into separate files, including one file containing GraphQL extend keyword on the other file's objects
-* Add of the _merge_ goal/task: it generates a GraphQL schema file, based on the source GraphQL schemas. It can be used to merge several GraphQL schema files into one file, or to reformat the schema files.
+* Add of the _merge_ goal/ _mergeTask_ task: it generates a GraphQL schema file, based on the source GraphQL schemas. It can be used to merge several GraphQL schema files into one file, or to reformat the schema files.
 
 Client mode:
 * Fixes #46 (logged in the GraphQL maven plugin project): Strings not properly escaped for JSON